requirements.txt
# Core dependencies
numpy
pandas
scipy
matplotlib

# Astronomy / cosmology
astropy

# API + HTTP handling
requests

# Jupyter + research workflow
jupyter
notebook

# Utilities
pyyaml
pyvis
markdown
pydantic
python-dotenv
PyYAML
python-dateutil
